qa-on-docs/requirements.txt
python-dotenv
langchain
openai
redis
tiktoken
autopep8